Enemy Tanks shoot and take damage
enemy tank will now aim towards the player and shoot in their direction. If the bullet hits the player, they lose health which is indicated in the top right corner. The player can also shoot and destroy the tank. At the moment, the game does not stop when the player dies, and the hitboxes need to be resized